Revert to legacy status shortcuts ruby script
